+++ AllaganLib.GameSheets/Sheets/Rows/ItemRow.cs
@@ -8,7 +8,9 @@ using AllaganLib.GameSheets.Model;
 using AllaganLib.GameSheets.Sheets.Helpers;
 using AllaganLib.Shared.Extensions;
 using AllaganLib.Shared.Time;
+using Lumina.Data.Parsing;
 using Lumina.Excel.Sheets;
+using LuminaSupplemental.Excel.Extensions;
 
 namespace AllaganLib.GameSheets.Sheets.Rows;
 
@@ -555,6 +557,82 @@ public partial class ItemRow : ExtendedRow<Item, ItemRow, ItemSheet>
             }
 
             return this.RowId.ToString();
+        }
+    }
+
+    // Shared Models
+
+    public List<ItemRow> GetSharedModels()
+    {
+        if (this.GetPrimaryModelKeyString() == string.Empty)
+        {
+            return new List<ItemRow>();
+        }
+
+        return this.Sheet.Where(c => c.GetPrimaryModelKeyString() != string.Empty && c.GetPrimaryModelKeyString() == this.GetPrimaryModelKeyString() && c.RowId != this.RowId).ToList();
+    }
+
+    public Quad GetPrimaryModelKey()
+    {
+        return (Quad)this.Base.ModelMain;
+    }
+
+    public string GetPrimaryModelKeyString()
+    {
+        var characterType = this.ModelCharacterType;
+        if (characterType != 0 && !ItemModelData.NoModelCategories.Contains(this.Base.ItemUICategory.RowId))
+        {
+            if (this.Base.Rarity != 7 && this.EquipSlotCategory != null)
+            {
+                var sEquipSlot = this.EquipSlotCategory.PossibleSlots.FirstOrNull();
+                if (sEquipSlot == null || !ItemModelData.ItemModelHelpers.TryGetValue((int)sEquipSlot, out var helper))
+                {
+                    return string.Empty;
+                }
+
+                if (helper == null)
+                {
+                    return string.Empty;
+                }
+
+                var key = this.GetPrimaryModelKey();
+                var modelKey = string.Format(helper.ModelFileFormat,key.A, key.B, key.C, key.D, characterType, (uint)sEquipSlot);
+                return modelKey;
+            }
+        }
+
+        return string.Empty;
+    }
+
+    public int ModelCharacterType
+    {
+        get
+        {
+            switch (this.Base.EquipRestriction)
+            {
+                case 0: return 0; // Not equippable
+                case 1: return 101; // Unrestricted, default to male hyur
+                case 2: return 101; // Any male
+                case 3: return 201; // Any female
+                case 4: return 101; // Hyur male
+                case 5: return 201; // Hyur female
+                case 6: return 501; // Elezen male
+                case 7: return 601; // Elezen female
+                case 8: return 1101; // Lalafell male
+                case 9: return 1201; // Lalafell female
+                case 10: return 701; // Miqo'te male
+                case 11: return 801; // Miqo'te female
+                case 12: return 901; // Roegadyn male
+                case 13: return 1001; // Roegadyn female
+                case 14: return 1301; // Au Ra male
+                case 15: return 1401; // Au Ra female
+                case 16: return 1501; // Hrothgar male
+                case 17: return 1801; // Viera female
+                case 18: return 1701; // Viera male
+                case 19: return 1601; // Hrothgar female
+            }
+
+            return 0;
         }
     }
 }